+++ components/Filter/index.js
@@ -1,11 +1,19 @@
 import React from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { addType, deleteType, initFeeds } from '../features/filteredPostSlice';
-import { View, Text, TouchableOpacity, StyleSheet } from 'react-native';
+import {
+  addType,
+  deleteType,
+  initFeeds
+} from '../../features/filteredPostSlice';
+import {
+  View,
+  Text,
+  TouchableOpacity
+} from 'react-native';
 
-import enumToString from '../constants/mapEnumToString';
-import hamsterTypes from '../constants/hamsterTypes';
-import colors from '../theme/color';
+import enumToString from '../../constants/mapEnumToString';
+import hamsterTypes from '../../constants/hamsterTypes';
+import styles from './styles';
 
 const Filter = ({ title }) => {
   const dispatch = useDispatch();
@@ -31,14 +39,14 @@ const Filter = ({ title }) => {
             <TouchableOpacity
               style={[
                 styles.tag,
-                { backgroundColor: selectedHamsterTypes[type] ? colors.main : colors.white }
+                selectedHamsterTypes[type] ? styles.mainBackground : styles.whiteBackground
               ]}
               key={type}
               onPress={() => handleSelectType(type)}
             >
               <Text style={[
                 styles.text,
-                { color: selectedHamsterTypes[type] ? colors.white : colors.main }
+                selectedHamsterTypes[type] ? styles.whiteText : styles.mainText
               ]}>{mapped[type]}</Text>
             </TouchableOpacity>
           );
@@ -47,32 +55,5 @@ const Filter = ({ title }) => {
     </View>
   );
 };
-
-const styles = StyleSheet.create({
-  container: {
-    margin: 10,
-    flexDirection: "column",
-    alignItems: "center",
-    backgroundColor: colors.white,
-    marginBottom: 0
-  },
-  tagContainer: {
-    flexDirection: "row",
-  },
-  tag: {
-    marginTop: 12,
-    width: 50,
-    height: 24,
-    margin: 5,
-    marginBottom: -20,
-    flexDirection: "row",
-    justifyContent: "center",
-    borderRadius: 6,
-  },
-  text: {
-    fontWeight: "bold",
-    alignSelf: "center",
-  }
-});
 
 export default Filter;
