+++ components/shared/Form/styles.js
@@ -1,0 +1,16 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  label: {
+    fontWeight: 'bold',
+    textAlign: 'center',
+    marginTop: 8,
+  },
+  button: {
+    width: '30%',
+    alignSelf: 'center',
+    margin: 30
+  }
+});
+
+export default styles;
