+++ components/shared/Input/styles.js
@@ -1,0 +1,15 @@
+import { StyleSheet } from 'react-native';
+import colors from '../../../theme/color';
+
+const styles = StyleSheet.create({
+  input: {
+    margin: 10,
+    padding: 13,
+    borderWidth: 1,
+    borderRadius: 5,
+    textAlign: 'center',
+    borderColor: colors.outline
+  },
+});
+
+export default styles;
