+++ components/shared/RadioButton/index.js
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
-import { View, Text, TouchableOpacity, StyleSheet } from 'react-native';
-import colors from '../../theme/color';
+import { View, Text, TouchableOpacity } from 'react-native';
+import styles from './styles';
 
 const RadioButton = ({ options, map, onChangeOption }) => {
   const [selectedOption, setSelectedOption] = useState();
@@ -28,33 +28,5 @@ const RadioButton = ({ options, map, onChangeOption }) => {
     </View>
   );
 };
-
-const styles = StyleSheet.create({
-  container: {
-    padding: 16,
-    margin: 9,
-    alignItems: 'center',
-    flexDirection: 'row',
-    justifyContent: 'space-evenly',
-    borderWidth: 1,
-    borderColor: colors.outline
-  },
-  option: {
-    padding: 8,
-    borderRadius: 20,
-    backgroundColor: colors.main
-  },
-  text: {
-    color: colors.white
-  },
-  selectedOption: {
-    padding: 8,
-    borderRadius: 20,
-    backgroundColor: colors.sub
-  },
-  selectedText: {
-    color: colors.black
-  },
-});
 
 export default RadioButton;
